Move side_panels.py to wagtail/admin/ui
