+++ wagtail/admin/ui/side_panels.py
